Fixed a corner case relating to screenshots.
When part of the game window was off-screen, the old screenshot code would not take into account the invisible area. The screenshots would come out with black space in the invisible area. This commit fixes that by creating a screen capture of the rectangle formed by the intersection of the game window bounds and the graphics device (i.e., screen) bounds.