Plugin manager changes.
Removed the function for stopping all the plugins. It is not that useful because that would realistically only be done when the entire application is shutting down. The function for starting all the plugins has been modified too. It now creates a dedicated thread to initialize and start each plugin. This is done to prevent the case when a plugin's initialize method blocks the current executing thread for a lengthy amount of time, preventing other plugins from being started.