Fixed a deadlock during plugin initialization.
If the implementation of `createMenuItem()` synchronized on the plugin object (e.g., had the `synchronized` modifier), then a deadlock would occur. This is because the initialization method calls `SwingUtilities.invokeAndWait(...)` which causes the Event Dispatch Thread wait on the thread that called `Plugin.initialize(...)` to return in order to obtain the lock. However, the thread that called `Plugin.initialize(...)` is also waiting for the Event Dispatch Thread to finish, so both threads end up waiting on each other indefinitely.